requirements.txt
aws-cdk-lib>=2.150.0
constructs>=10.3.0
cdk-nag>=2.28.158
boto3>=1.34.150
